pyproject.toml
[project]
name = "pythonsqlalchemydemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
